pyproject.toml
[project]
name = "gemini-code-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "google-genai>=1.24.0",
    "pydantic-settings>=2.10.1",
    "pygithub>=2.6.1",
    "ruff>=0.12.2",
]
